+++ client/jetbrains/webview/src/bridge-mock/index.ts
@@ -1,17 +1,23 @@
 import { SearchPatternType } from '@sourcegraph/shared/src/graphql-operations'
 
-import type { Request } from '../search/js-to-java-bridge'
+import type { PreviewRequest, Request } from '../search/js-to-java-bridge'
 import type { Search } from '../search/types'
 
-let savedSearch: Search = {
-    query: 'r:github.com/sourcegraph/sourcegraph jetbrains',
-    caseSensitive: false,
-    patternType: SearchPatternType.literal,
-    selectedSearchContextSpec: 'global',
-}
+const instanceURL = 'https://sourcegraph.com'
 
 const codeDetailsNode = document.querySelector('#code-details') as HTMLPreElement
 const iframeNode = document.querySelector('#webview') as HTMLIFrameElement
+
+const savedSearchFromLocalStorage = localStorage.getItem('savedSearch')
+let savedSearch: Search = savedSearchFromLocalStorage
+    ? (JSON.parse(savedSearchFromLocalStorage) as Search)
+    : {
+          query: 'r:github.com/sourcegraph/sourcegraph jetbrains',
+          caseSensitive: false,
+          patternType: SearchPatternType.literal,
+          selectedSearchContextSpec: 'global',
+      }
+let previewContent: PreviewRequest['arguments'] | null = null
 
 function callJava(request: Request): Promise<object> {
     return new Promise((resolve, reject) => {
@@ -37,7 +43,7 @@ function handleRequest(
         case 'getConfig': {
             onSuccessCallback(
                 JSON.stringify({
-                    instanceURL: 'https://sourcegraph.com',
+                    instanceURL,
                     isGlobbingEnabled: true,
                     accessToken: null,
                 })
@@ -62,16 +68,18 @@ function handleRequest(
         }
 
         case 'preview': {
-            const { content, absoluteOffsetAndLengths } = request.arguments
+            previewContent = request.arguments
 
-            const start = absoluteOffsetAndLengths.length > 0 ? absoluteOffsetAndLengths[0][0] : 0
-            const length = absoluteOffsetAndLengths.length > 0 ? absoluteOffsetAndLengths[0][1] : 0
+            const start =
+                previewContent.absoluteOffsetAndLengths.length > 0 ? previewContent.absoluteOffsetAndLengths[0][0] : 0
+            const length =
+                previewContent.absoluteOffsetAndLengths.length > 0 ? previewContent.absoluteOffsetAndLengths[0][1] : 0
 
             let htmlContent: string
-            if (content === null) {
+            if (previewContent.content === null) {
                 htmlContent = '(No preview available)'
             } else {
-                const decodedContent = atob(content)
+                const decodedContent = atob(previewContent.content)
                 htmlContent = escapeHTML(decodedContent.slice(0, start))
                 htmlContent += `<span id="code-details-highlight">${escapeHTML(
                     decodedContent.slice(start, start + length)
@@ -94,14 +102,19 @@ function handleRequest(
         }
 
         case 'open': {
-            const { path } = request.arguments
-            alert(`Opening ${path}`)
+            previewContent = request.arguments
+            if (previewContent.fileName) {
+                alert(`Now the IDE would open ${previewContent.path} in the editor...`)
+            } else {
+                window.open(instanceURL + previewContent.relativeUrl, '_blank')
+            }
             onSuccessCallback('null')
             break
         }
 
         case 'saveLastSearch': {
             savedSearch = request.arguments
+            localStorage.setItem('savedSearch', JSON.stringify(savedSearch))
             onSuccessCallback('null')
             break
         }
@@ -128,7 +141,10 @@ iframeNode.addEventListener('load', () => {
     const iframeWindow = iframeNode.contentWindow
     if (iframeWindow !== null) {
         iframeWindow.callJava = callJava
-        iframeWindow.initializeSourcegraph()
+        iframeWindow
+            .initializeSourcegraph()
+            .then(() => {})
+            .catch(() => {})
     }
 })