+++ client/web/src/auth/SignupEmailField.tsx
@@ -4,7 +4,7 @@ import classNames from 'classnames'
 
 import { LoaderInput } from '@sourcegraph/branded/src/components/LoaderInput'
 import { deriveInputClassName, InputValidationState } from '@sourcegraph/shared/src/util/useInputValidation'
-import { Typography } from '@sourcegraph/wildcard'
+import { Label } from '@sourcegraph/wildcard'
 
 import { EmailInput } from './SignInSignUpCommon'
 
@@ -24,14 +24,14 @@ export const SignupEmailField: React.FunctionComponent<React.PropsWithChildren<S
     emailInputReference,
 }) => (
     <div className="form-group d-flex flex-column align-content-start">
-        <Typography.Label
+        <Label
             htmlFor="email"
             className={classNames('align-self-start', {
                 'text-danger font-weight-bold': emailState.kind === 'INVALID',
             })}
         >
             {label}
-        </Typography.Label>
+        </Label>
         <LoaderInput className={deriveInputClassName(emailState)} loading={emailState.kind === 'LOADING'}>
             <EmailInput
                 className={deriveInputClassName(emailState)}
@@ -42,8 +42,13 @@ export const SignupEmailField: React.FunctionComponent<React.PropsWithChildren<S
                 autoFocus={true}
                 placeholder=" "
                 inputRef={emailInputReference}
+                aria-describedby="email-input-invalid-feedback"
             />
         </LoaderInput>
-        {emailState.kind === 'INVALID' && <small className="invalid-feedback">{emailState.reason}</small>}
+        {emailState.kind === 'INVALID' && (
+            <small className="invalid-feedback" id="email-input-invalid-feedback">
+                {emailState.reason}
+            </small>
+        )}
     </div>
 )