+++ client/web/src/org/members/OrgPendingInvites.tsx
@@ -19,7 +19,7 @@ import {
     MenuItem,
     Position,
     PageSelector,
-    Typography,
+    H3,
 } from '@sourcegraph/wildcard'
 
 import { PageTitle } from '../../components/PageTitle'
@@ -377,7 +377,7 @@ export const OrgPendingInvitesPage: React.FunctionComponent<React.PropsWithChild
                 {authenticatedUser && data && data.pendingInvitations.length === 0 && (
                     <Container>
                         <div className="d-flex flex-0 flex-column justify-content-center align-items-center">
-                            <Typography.H3>No invites pending</Typography.H3>
+                            <H3>No invites pending</H3>
                             <div>
                                 <InviteMemberModalHandler
                                     orgName={org.name}