+++ dev/sg/sg_rfc.go
@@ -34,21 +34,21 @@ func rfcExec(ctx context.Context, args []string) error {
 
 	switch args[0] {
 	case "list":
-		return rfc.List(ctx, std.Out.Output)
+		return rfc.List(ctx, std.Out)
 
 	case "search":
 		if len(args) != 2 {
 			return errors.New("no search query given")
 		}
 
-		return rfc.Search(ctx, args[1], std.Out.Output)
+		return rfc.Search(ctx, args[1], std.Out)
 
 	case "open":
 		if len(args) != 2 {
 			return errors.New("no number given")
 		}
 
-		return rfc.Open(ctx, args[1], std.Out.Output)
+		return rfc.Open(ctx, args[1], std.Out)
 	}
 
 	return nil