+++ internal/cmd/tracking-issue/main.go
@@ -10,9 +10,8 @@ import (
 	"strings"
 
 	"github.com/machinebox/graphql"
-	"golang.org/x/oauth2"
-
 	"github.com/sourcegraph/sourcegraph/lib/errors"
+	"golang.org/x/oauth2"
 )
 
 const (
@@ -32,8 +31,21 @@ func main() {
 	flag.Parse()
 
 	if err := run(*token, *org, *dry, *verbose); err != nil {
-		log.Fatal(err)
+		if isRateLimitErr(err) {
+			log.Printf("Github API limit reached - soft failing. Err: %s\n", err)
+		} else {
+			log.Fatal(err)
+		}
 	}
+}
+
+func isRateLimitErr(err error) bool {
+	if err == nil {
+		return false
+	}
+
+	baseErr := errors.UnwrapAll(err)
+	return strings.Contains(baseErr.Error(), "API rate limit exceeded")
 }
 
 func run(token, org string, dry, verbose bool) (err error) {